Greatly cleanup and works towards ESM (#32)
* Revert "Add Vitest for unit tests (#30)"

This reverts commit 52aa813616d3b69dfb3abf466208730392c81ec6.

* Revert "Remove dupe config file (#31)"

This reverts commit aae4f9e45fda8447df3f8a52143aaa5286425cce.

* Remove rootDirs from client

* Remove rootDirs from server

* Add compile-ts to client

* Resolve modules via NodeNext

* Make client an ESM module

* Add compile-ts to server

* Update module resolution of server

* Make server an ESM module

* Use bundler resolution in client

* Remove unused tsconfig.json

* Use NodeNext for compatibility with AHK++

* Add logs

* Progress

* ESM but still not quite working

* More CJS building

* Fix clean

* Use server.cjs

* Further cleanup

* Ignore extension.cts when prettifying
