+++ .gitignore
@@ -5,9 +5,7 @@ node_modules
 *.vsix
 scripts/*
 *.js
+*.cjs
 *.js.map
 *.tsbuildinfo
 **/dist/**/*.d.ts
-
-# Vitest
-coverage
