+++ .vscode/launch.json
@@ -34,7 +34,7 @@
 			"runtimeExecutable": "${execPath}",
 			"args": ["--extensionDevelopmentPath=${workspaceFolder}"],
 			"env": {
-				"VSCODE_AHK_SERVER_PATH": "server/dist/server.js",
+				"VSCODE_AHK_SERVER_PATH": "server/dist/server.cjs",
 				"SYNTAXES_PATH": "syntaxes"
 			},
 			"outFiles": ["${workspaceFolder}/client/dist/**/*.js"],
