+++ client/package.json
@@ -3,6 +3,10 @@
 	"description": "VSCode part of a language server",
 	"license": "MIT",
 	"version": "0.0.1",
+	"type": "module",
+	"scripts": {
+		"compile-ts": "tsc"
+	},
 	"engines": {
 		"vscode": "^1.43.0"
 	}
