+++ client/src/extension.cts
@@ -29,17 +29,17 @@ import {
 	LanguageClientOptions,
 	ServerOptions,
 	TransportKind
-} from 'vscode-languageclient/node';
+} from 'vscode-languageclient/node.js';
 import { resolve } from 'path';
 import { ChildProcess, execSync, spawn } from 'child_process';
 import { readdirSync, readFileSync, lstatSync, readlinkSync, unlinkSync, writeFileSync } from 'fs';
-import { CfgKey, getAhkppConfig, getCfg, ShowOutput } from './config';
-import { resolvePath } from './utils';
+import { CfgKey, getAhkppConfig, getCfg, ShowOutput } from './config.cjs';
 
 let client: LanguageClient, outputchannel: OutputChannel, ahkStatusBarItem: StatusBarItem;
 const ahkprocesses = new Map<number, ChildProcess & { path?: string }>();
 let v2Interpreter = getCfg<string>(CfgKey.InterpreterPathV2), server_is_ready = false
 const textdecoders: TextDecoder[] = [new TextDecoder('utf8', { fatal: true }), new TextDecoder('utf-16le', { fatal: true })];
+const isWindows = process.platform === 'win32';
 let extlist: string[] = [], debugexts: Record<string, string> = {}, langs: string[] = [];
 const loadedCollection = {
 	'ahk2.browse': 'Browse your file system to find AutoHotkey2 interpreter',
@@ -61,8 +61,11 @@ const loadedCollection = {
 };
 
 export function activate(context: ExtensionContext): Promise<LanguageClient> {
+	// when debugging, this goes into the Debug Console (Ctrl + Shift + Y)
+	console.log('Activating AHK v2 language server');
+
 	/** Absolute path to `server.js` */
-	const defaultServerModule = context.asAbsolutePath(`ahk2/server/dist/server.js`);
+	const defaultServerModule = context.asAbsolutePath(`ahk2/server/dist/server.cjs`);
 	const serverModule = process.env.VSCODE_AHK_SERVER_PATH ? context.asAbsolutePath(process.env.VSCODE_AHK_SERVER_PATH) : defaultServerModule;
 
 	// If the extension is launched in debug mode then the debug server options are used
@@ -621,6 +624,34 @@ async function onDidChangegetInterpreter() {
 		ahkStatusBarItem.text = localize('ahk2.set.interpreter')
 		ahkStatusBarItem.tooltip = undefined, ahkPath = '';
 	}
+}
+
+/**
+ * Returns the provided path as an absolute path.
+ * Resolves the provided path against the provided workspace.
+ * Resolves symbolic links by default.
+ * Returns empty string if resolution fails.
+ */
+export function resolvePath(path: string | undefined, workspace?: string, resolveSymbolicLink = true): string {
+	if (!path)
+		return '';
+	const paths: string[] = [];
+	// If the path does not contain a colon, resolve it relative to the workspace
+	if (!path.includes(':'))
+		paths.push(resolve(workspace ?? '', path));
+	// If there are no slashes or backslashes in the path and the platform is Windows
+	if (!/[\\/]/.test(path) && isWindows)
+		paths.push(execSync(`where ${path}`, { encoding: 'utf-8' }).trim());
+	paths.push(path);
+	for (let path of paths) {
+		if (!path) continue;
+		try {
+			if (lstatSync(path).isSymbolicLink() && resolveSymbolicLink)
+				path = resolve(path, '..', readlinkSync(path));
+			return path;
+		} catch { }
+	}
+	return '';
 }
 
 /**
