+++ server/package.json
@@ -2,6 +2,10 @@
 	"name": "vscode-autohotkey2-server",
 	"description": "Autohotkey2 Language Server using vscode-lsp.",
 	"license": "LGPLv3.0",
+	"type": "module",
+	"scripts": {
+		"compile-ts": "tsc"
+	},
 	"engines": {
 		"node": "*"
 	},
