+++ server/src/common.ts
@@ -4,25 +4,25 @@ import { readdirSync, readFileSync, existsSync, statSync, promises as fs } from 
 import { Connection, MessageConnection } from 'vscode-languageserver';
 import { TextDocument } from 'vscode-languageserver-textdocument';
 import { CompletionItem, CompletionItemKind, Hover, InsertTextFormat, Range, SymbolKind } from 'vscode-languageserver-types';
-import { AhkSymbol, Lexer, setCommentTagRegex } from './Lexer';
-import { diagnostic } from './localize';
-import { isBrowser, jsDocTagNames } from './constants';
-import { AhkppConfig, CfgKey, getCfg, newAhkppConfig } from './config';
-export * from './codeActionProvider';
-export * from './colorProvider';
-export * from './commandProvider';
-export * from './completionProvider';
-export * from './constants';
-export * from './definitionProvider';
-export * from './formattingProvider';
-export * from './hoverProvider';
-export * from './Lexer';
-export * from './localize';
-export * from './referencesProvider';
-export * from './renameProvider';
-export * from './semanticTokensProvider';
-export * from './signatureProvider';
-export * from './symbolProvider';
+import { AhkSymbol, Lexer, setCommentTagRegex } from './Lexer.js';
+import { diagnostic } from './localize.js';
+import { isBrowser, jsDocTagNames } from './constants.js';
+import { AhkppConfig, CfgKey, getCfg, newAhkppConfig } from './config.js';
+export * from './codeActionProvider.js';
+export * from './colorProvider.js';
+export * from './commandProvider.js';
+export * from './completionProvider.js';
+export * from './constants.js';
+export * from './definitionProvider.js';
+export * from './formattingProvider.js';
+export * from './hoverProvider.js';
+export * from './Lexer.js';
+export * from './localize.js';
+export * from './referencesProvider.js';
+export * from './renameProvider.js';
+export * from './semanticTokensProvider.js';
+export * from './signatureProvider.js';
+export * from './symbolProvider.js';
 
 export const winapis: string[] = [];
 export const lexers: Record<string, Lexer> = {};
