+++ server/src/hoverProvider.ts
@@ -2,7 +2,7 @@ import { CancellationToken, Hover, HoverParams, SymbolKind } from 'vscode-langua
 import {
 	AhkSymbol, FuncNode, Maybe, SemanticTokenTypes, Variable,
 	get_detail, hoverCache, join_types, lexers, find_symbols
-} from './common';
+} from './common.js';
 
 export async function hoverProvider(params: HoverParams, token: CancellationToken): Promise<Maybe<Hover>> {
 	if (token.isCancellationRequested) return;
