+++ server/src/referencesProvider.ts
@@ -2,7 +2,7 @@ import { CancellationToken, Location, Range, ReferenceParams, SymbolKind } from 
 import {
 	AhkSymbol, ClassNode, Context, FuncNode, FuncScope, Lexer, Property, Variable,
 	ahkuris, ahkvars, find_symbol, find_symbols, lexers
-} from './common';
+} from './common.js';
 
 export async function referenceProvider(params: ReferenceParams, token: CancellationToken): Promise<Location[] | undefined> {
 	const result: Location[] = [], doc = lexers[params.textDocument.uri.toLowerCase()];
