--- tsconfig.json
@@ -1,14 +1,0 @@
-{
-	"compilerOptions": {
-		"module": "commonjs",
-		"target": "ES2021",
-		"lib": ["ES2021"],
-		"outDir": "out",
-		"rootDir": "src",
-		"sourceMap": true,
-		"composite": true
-	},
-	"include": ["src"],
-	"exclude": ["node_modules", ".vscode-test"],
-	"references": [{ "path": "./client" }, { "path": "./server" }]
-}
