--- vitest.config.mjs
@@ -1,13 +1,0 @@
-import { defineConfig } from 'vitest/config';
-
-// https://vitest.dev/config
-// https://vitest.dev/guide/improving-performance.html
-export default defineConfig({
-	test: {
-		isolate: false,
-		pool: 'threads',
-		coverage: {
-			enabled: true,
-		},
-	},
-});
